+++ application/apps/indexer/sources/src/socket/tcp.rs
@@ -1,10 +1,15 @@
+use std::time::Duration;
+
 use crate::{
     socket::ReconnectStateMsg, ByteSource, Error as SourceError, ReloadInfo, SourceFilter,
 };
 use bufread::DeqBuffer;
-use tokio::{net::TcpStream, task::yield_now};
+use tokio::{net::TcpStream, task::yield_now, time::timeout};
 
-use super::{ReconnectInfo, ReconnectResult, ReconnectToServer, MAX_BUFF_SIZE, MAX_DATAGRAM_SIZE};
+use super::{
+    hanlde_buff_capacity, BuffCapacityState, ReconnectInfo, ReconnectResult, ReconnectToServer,
+    MAX_BUFF_SIZE, MAX_DATAGRAM_SIZE,
+};
 
 pub struct TcpSource {
     buffer: DeqBuffer,
@@ -93,18 +98,41 @@ impl ByteSource for TcpSource {
         // If buffer is almost full then skip loading and return the available bytes.
         // This can happen because some parsers will parse the first item of the provided slice
         // while the producer will call load on each iteration making data accumulate.
-        if self.buffer.write_available() < MAX_DATAGRAM_SIZE {
-            let available_bytes = self.len();
-            return Ok(Some(ReloadInfo::new(0, available_bytes, 0, None)));
+        match hanlde_buff_capacity(&mut self.buffer) {
+            BuffCapacityState::CanLoad => {}
+            BuffCapacityState::AlmostFull => {
+                let available_bytes = self.len();
+                return Ok(Some(ReloadInfo::new(0, available_bytes, 0, None)));
+            }
         }
 
         // TODO use filter
         loop {
             debug!("Wait for tcp socket to become readable");
-            self.socket
-                .readable()
-                .await
-                .map_err(|e| SourceError::Unrecoverable(format!("{e}")))?;
+
+            if self.reconnect_info.is_some() {
+                // It's possible for readable() message to stay stuck in case of server reset.
+                // To avoid that we try to reconnect after 2 seconds without having data ready
+                // to be written, assuming the server is down.
+                match timeout(Duration::from_secs(2), self.socket.readable()).await {
+                    Ok(Ok(())) => {}
+                    Ok(Err(err)) => return Err(SourceError::Unrecoverable(format!("{err}"))),
+                    Err(_) => {
+                        match self.reconnect().await {
+                            ReconnectResult::Reconnected => {}
+                            ReconnectResult::NotConfigured => {}
+                            ReconnectResult::Error(error) => {
+                                return Err(SourceError::Unrecoverable(error.to_string()));
+                            }
+                        };
+                    }
+                };
+            } else {
+                self.socket
+                    .readable()
+                    .await
+                    .map_err(|e| SourceError::Unrecoverable(e.to_string()))?;
+            }
             debug!("Socket ready to read");
             match self.socket.try_read(&mut self.tmp_buffer) {
                 Ok(len) => {
@@ -167,11 +195,6 @@ impl ByteSource for TcpSource {
 
     fn consume(&mut self, offset: usize) {
         self.buffer.read_done(offset);
-        // Calling read_done() won't make free up writable memory.
-        // Therefore we need to call flush manually.
-        if self.buffer.write_available() < MAX_DATAGRAM_SIZE {
-            self.buffer.flush();
-        }
     }
 
     fn len(&self) -> usize {